backend/requirements.txt
Django==4.2
djangorestframework==3.14
psycopg2-binary
python-decouple
python-dotenv
django-cors-headers
dj-database-url>=2.0
requests
pydantic==2
gunicorn==21
celery[redis]==5.3
redis>=5.0
openai>=1.0
